+++ src/Symfony/Component/HttpFoundation/IpUtils.php
@@ -35,6 +35,8 @@ class IpUtils
     public static function checkIp(?string $requestIp, string|array $ips): bool
     {
         if (null === $requestIp) {
+            trigger_deprecation('symfony/http-foundation', '5.4', 'Passing null as $requestIp to "%s()" is deprecated, pass an empty string instead.', __METHOD__);
+
             return false;
         }
 
@@ -63,6 +65,12 @@ class IpUtils
      */
     public static function checkIp4(?string $requestIp, string $ip): bool
     {
+        if (null === $requestIp) {
+            trigger_deprecation('symfony/http-foundation', '5.4', 'Passing null as $requestIp to "%s()" is deprecated, pass an empty string instead.', __METHOD__);
+
+            return false;
+        }
+
         $cacheKey = $requestIp.'-'.$ip;
         if (isset(self::$checkedIps[$cacheKey])) {
             return self::$checkedIps[$cacheKey];
@@ -108,6 +116,12 @@ class IpUtils
      */
     public static function checkIp6(?string $requestIp, string $ip): bool
     {
+        if (null === $requestIp) {
+            trigger_deprecation('symfony/http-foundation', '5.4', 'Passing null as $requestIp to "%s()" is deprecated, pass an empty string instead.', __METHOD__);
+
+            return false;
+        }
+
         $cacheKey = $requestIp.'-'.$ip;
         if (isset(self::$checkedIps[$cacheKey])) {
             return self::$checkedIps[$cacheKey];